+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Added
 - You can now edit an already scanned item's price from the cart.
 - You can now search for customers via name, email, telephone rather than having to enter the full email address.
-- A POS order billing address will now be the customers default billing address or the physical store's address (pulled from core config).
+- A POS order billing address can now be the customers default billing address or the physical store's address (pulled from core config).
 - Added the optional ability to emulate the physical store's address as the POS order's shipping address, despite the order being virtual and not having one.
 
 ### Changed